pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spurs-survey"
version = "0.1.0"
description = "Automated post-match rating surveys for r/coys"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "Pillow>=10.0",
    "thefuzz[speedup]>=0.20",
    "google-api-python-client>=2.100",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.90",
    "pytest>=7.4",
]

[project.scripts]
spurs-survey = "spurs_survey.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
